replaced update method with merge in feature_template_field
