+++ vmngclient/utils/dict.py
@@ -1,0 +1,14 @@
+def merge(a, b, path=None):
+    if path is None:
+        path = []
+    for key in b:
+        if key in a:
+            if isinstance(a[key], dict) and isinstance(b[key], dict):
+                merge(a[key], b[key], path + [str(key)])
+            elif a[key] == b[key]:
+                pass  # same leaf value
+            else:
+                raise Exception(f"Conflict at {'.'.join(path + [str(key)])}")
+        else:
+            a[key] = b[key]
+    return a
